Refactor main callbacks and handlers to use roles
