+++ app/roles/admin/tags_management/handlers/tags_create.py
@@ -4,11 +4,14 @@ from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import Message, CallbackQuery
 
 from app.database.tag_db_operations import add_tag_to_db
+from app.database.user_db_operations import get_user_by_telegram_tag
 from app.keyboards import (
-    inline_single_cancel_button, main_actions_keyboard, )
+    inline_single_cancel_button,
+    main_actions_keyboard,
+)
+from app.roles.admin.admin import admin
 from app.roles.user.callbacks_enum import Callbacks
-from app.roles.user.main_actions.tags_management.handlers.tags_read import manage_tags
-from app.roles.user.user_cmds import user
+from app.roles.admin.tags_management.handlers.tags_read import manage_tags
 
 
 class TagCreationStates(StatesGroup):
@@ -26,14 +29,18 @@ async def on_create_tag_clicked(event: Message | CallbackQuery, state: FSMContex
     await state.set_state(TagCreationStates.waiting_for_tag_name)
 
 
-@user.callback_query(F.data == Callbacks.tag_create_callback)
+@admin.callback_query(F.data == Callbacks.tag_create_callback)
 async def handle_create_tag_callback(callback: CallbackQuery, state: FSMContext):
     await on_create_tag_clicked(callback, state)
 
 
-@user.message(TagCreationStates.waiting_for_tag_name)
+@admin.message(TagCreationStates.waiting_for_tag_name)
 async def process_tag_name(message: Message, state: FSMContext):
     tag_name = message.text.strip()
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     if len(tag_name) > 32:
         await message.answer(
             text="Название тега слишком длинное! Максимум 32 символа. \nВведите новое название:",
@@ -44,7 +51,7 @@ async def process_tag_name(message: Message, state: FSMContext):
     if success:
         await message.answer(
             text=response,
-            reply_markup=main_actions_keyboard
+            reply_markup=main_actions_keyboard(user.role)
         )
         await state.clear()
     else:
@@ -54,6 +61,6 @@ async def process_tag_name(message: Message, state: FSMContext):
         )
 
 
-@user.callback_query(F.data == Callbacks.cancel_tag_naming_callback)
+@admin.callback_query(F.data == Callbacks.cancel_tag_naming_callback)
 async def on_cancel_tag_naming(callback: CallbackQuery, state: FSMContext):
     await manage_tags(callback, state)
