+++ app/roles/owner/admins_management/admins_management.py
@@ -14,14 +14,15 @@ from app.database.user_db_operations import (
     add_or_update_user_to_admin,
     demote_admin_to_user,
     get_user_by_id,
+    get_user_by_telegram_tag,
 )
 from app.keyboards import (
     inline_admin_list,
     inline_single_cancel_button,
     main_actions_keyboard,
 )
+from app.roles.owner.owner import owner  # Импортируем роутер owner
 from app.roles.user.callbacks_enum import Callbacks
-from app.roles.user.user_cmds import user
 from app.utils.logger import logger
 
 
@@ -30,14 +31,6 @@ class AdminManagementStates(StatesGroup):
 
 
 def is_valid_telegram_username(username: str) -> bool:
-    """
-    Validates Telegram-username.
-    Username must:
-    - Start with @
-    - Have 5 or more symbols
-    - No digit or _ after @
-    - Have only letters (a-z, A-Z), digits (0-9) and _
-    """
     return (
             username.startswith("@")
             and len(username) >= 5
@@ -47,9 +40,13 @@ def is_valid_telegram_username(username: str) -> bool:
     )
 
 
-@user.message(F.text == labels.MANAGE_ADMINS)
+@owner.message(F.text == labels.MANAGE_ADMINS)
 async def manage_admins(message: Message):
     logger.info("manage_admins_call")
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     await message.answer(
         text="Выберите админа или создайте нового:",
         reply_markup=await inline_admin_list(
@@ -66,12 +63,12 @@ async def on_add_admin_clicked(callback: CallbackQuery, state: FSMContext):
     await callback.answer()
 
 
-@user.callback_query(F.data == Callbacks.add_admin_callback)
+@owner.callback_query(F.data == Callbacks.add_admin_callback)
 async def handle_add_admin_callback(callback: CallbackQuery, state: FSMContext):
     await on_add_admin_clicked(callback, state)
 
 
-@user.message(AdminManagementStates.waiting_for_admin_username)
+@owner.message(AdminManagementStates.waiting_for_admin_username)
 async def process_admin_username(message: Message, state: FSMContext):
     username = message.text.strip()
     if not is_valid_telegram_username(username):
@@ -91,8 +88,12 @@ async def process_admin_username(message: Message, state: FSMContext):
         )
         return
     success, response = await add_or_update_user_to_admin(username)
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     if success:
-        await message.answer(text=response, reply_markup=main_actions_keyboard)
+        await message.answer(text=response, reply_markup=main_actions_keyboard(user.role))
         await state.clear()
     else:
         await message.answer(
@@ -104,7 +105,7 @@ async def process_admin_username(message: Message, state: FSMContext):
     await message.bot.delete_message(chat_id=message.chat.id, message_id=message.message_id - 1)
 
 
-@user.callback_query(F.data.startswith("admin_clicked"))
+@owner.callback_query(F.data.startswith("admin_clicked"))
 async def on_admin_clicked(callback: CallbackQuery):
     try:
         user_id = callback.data.split(":")[1]
@@ -140,7 +141,7 @@ async def on_admin_clicked(callback: CallbackQuery):
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith(Callbacks.admin_delete_callback))
+@owner.callback_query(F.data.startswith(Callbacks.admin_delete_callback))
 async def on_admin_delete_callback(callback: CallbackQuery):
     try:
         user_id = callback.data.split(":")[1]
@@ -148,12 +149,16 @@ async def on_admin_delete_callback(callback: CallbackQuery):
         await callback.answer("Ошибка: пользователь не выбран!", show_alert=True)
         return
     success, response = await demote_admin_to_user(user_id)
-    await callback.message.answer(text=response, reply_markup=main_actions_keyboard)
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
+    await callback.message.answer(text=response, reply_markup=main_actions_keyboard(user.role))
     await callback.message.delete()
     await callback.answer("")
 
 
-@user.callback_query(F.data == Callbacks.return_to_admin_list_callback)
+@owner.callback_query(F.data == Callbacks.return_to_admin_list_callback)
 async def on_return_to_admin_list(callback: CallbackQuery):
     await callback.message.edit_text(
         text="Выберите админа или создайте нового:",
@@ -162,3 +167,18 @@ async def on_return_to_admin_list(callback: CallbackQuery):
         ),
     )
     await callback.answer("")
+
+
+@owner.callback_query(F.data == Callbacks.cancel_primary_action_callback)
+async def on_cancel_primary_callback(callback: CallbackQuery, state: FSMContext):
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
+    await callback.answer("")
+    await callback.message.delete()
+    await callback.message.answer(
+        text="Действие отменено. Выберите новое действие с помощью кнопок под клавиатурой.",
+        reply_markup=main_actions_keyboard(user.role),
+    )
+    await state.clear()
