+++ app/roles/user/main_actions/recording_create/recording_create.py
@@ -9,14 +9,15 @@ from bson import ObjectId
 from app.config import labels
 from app.database.conference_db_operations import add_conference_to_db, conference_exists_by_link
 from app.database.tag_db_operations import get_tag_by_id
+from app.database.user_db_operations import get_user_by_telegram_tag
 from app.keyboards import (
     inline_active_tag_list,
     inline_single_cancel_button,
     main_actions_keyboard,
 )
+from app.roles.admin.admin import admin  # Импортируем роутер admin
 from app.roles.user.callbacks_enum import Callbacks
 from app.utils.logger import logger
-from app.roles.user.user_cmds import user
 
 
 class RecordingCreateStates(StatesGroup):
@@ -28,9 +29,8 @@ class RecordingCreateStates(StatesGroup):
     waiting_for_periodicity = State()
 
 
-@user.message(F.text == labels.RECORD)
+@admin.message(F.text == labels.RECORD)
 async def start_recording(message: Message, state: FSMContext):
-    """Start the recording creation process by selecting a tag."""
     logger.info("start_recording_call")
     await message.answer(
         text="Выберите тег для новой конференции:",
@@ -44,9 +44,8 @@ async def start_recording(message: Message, state: FSMContext):
     await state.set_state(RecordingCreateStates.waiting_for_tag)
 
 
-@user.callback_query(F.data.startswith(Callbacks.tag_clicked_in_recording_mode_callback))
+@admin.callback_query(F.data.startswith(Callbacks.tag_clicked_in_recording_mode_callback))
 async def process_tag_for_recording(callback: CallbackQuery, state: FSMContext):
-    """Process the selected tag and ask for the conference link."""
     try:
         tag_id = callback.data.split(":")[1]
     except IndexError:
@@ -70,10 +69,9 @@ async def process_tag_for_recording(callback: CallbackQuery, state: FSMContext):
     await callback.answer("")
 
 
-@user.callback_query(F.data == Callbacks.back_to_tag_in_create_conference_mode,
-                     RecordingCreateStates.waiting_for_meet_link)
+@admin.callback_query(F.data == Callbacks.back_to_tag_in_create_conference_mode,
+                      RecordingCreateStates.waiting_for_meet_link)
 async def back_to_tag_from_link(callback: CallbackQuery, state: FSMContext):
-    """Return to tag selection from link input."""
     await callback.message.edit_text(
         text="Выберите тег для новой конференции:",
         reply_markup=await inline_active_tag_list(
@@ -87,17 +85,20 @@ async def back_to_tag_from_link(callback: CallbackQuery, state: FSMContext):
     await callback.answer("")
 
 
-@user.message(RecordingCreateStates.waiting_for_meet_link)
+@admin.message(RecordingCreateStates.waiting_for_meet_link)
 async def process_meet_link_for_recording(message: Message, state: FSMContext):
-    """Process the conference link, validate it as Google Meet, and ask for the timezone."""
     meet_link = message.text.strip()
     state_data = await state.get_data()
     tag_id = state_data.get("tag_id")
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
 
     if not tag_id:
         await message.answer(
             text="Ошибка: тег не выбран! Попробуйте начать заново.",
-            reply_markup=main_actions_keyboard,
+            reply_markup=main_actions_keyboard(user.role),
         )
         await state.clear()
         return
@@ -128,9 +129,8 @@ async def process_meet_link_for_recording(message: Message, state: FSMContext):
     await state.set_state(RecordingCreateStates.waiting_for_timezone)
 
 
-@user.callback_query(F.data == "back_to_link", RecordingCreateStates.waiting_for_timezone)
+@admin.callback_query(F.data == "back_to_link", RecordingCreateStates.waiting_for_timezone)
 async def back_to_link_from_timezone(callback: CallbackQuery, state: FSMContext):
-    """Return to link input from timezone."""
     await callback.message.edit_text(
         text="Введите ссылку на Google Meet конференцию:",
         reply_markup=InlineKeyboardMarkup(inline_keyboard=[
@@ -141,13 +141,16 @@ async def back_to_link_from_timezone(callback: CallbackQuery, state: FSMContext)
     await callback.answer("")
 
 
-@user.message(RecordingCreateStates.waiting_for_timezone)
+@admin.message(RecordingCreateStates.waiting_for_timezone)
 async def process_timezone(message: Message, state: FSMContext):
-    """Process the timezone and ask for the start date."""
     timezone_str = message.text.strip()
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     try:
-        timezone = int(timezone_str)  # Expecting something like +3 or -5
-        if not -12 <= timezone <= 14:  # Reasonable timezone range
+        timezone = int(timezone_str)
+        if not -12 <= timezone <= 14:
             raise ValueError
     except ValueError:
         await message.answer(
@@ -168,9 +171,8 @@ async def process_timezone(message: Message, state: FSMContext):
     await state.set_state(RecordingCreateStates.waiting_for_start_date)
 
 
-@user.callback_query(F.data == "back_to_timezone", RecordingCreateStates.waiting_for_start_date)
+@admin.callback_query(F.data == "back_to_timezone", RecordingCreateStates.waiting_for_start_date)
 async def back_to_timezone_from_date(callback: CallbackQuery, state: FSMContext):
-    """Return to timezone input from start date."""
     await callback.message.edit_text(
         text="Укажите тайм-зону относительно UTC (например, +3 или -5):",
         reply_markup=InlineKeyboardMarkup(inline_keyboard=[
@@ -181,20 +183,21 @@ async def back_to_timezone_from_date(callback: CallbackQuery, state: FSMContext)
     await callback.answer("")
 
 
-@user.message(RecordingCreateStates.waiting_for_start_date)
+@admin.message(RecordingCreateStates.waiting_for_start_date)
 async def process_start_date(message: Message, state: FSMContext):
-    """Process the start date, validate it, and ask about recurrence."""
     start_date_str = message.text.strip()
     state_data = await state.get_data()
     timezone = state_data.get("timezone")
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
 
     try:
         start_date = datetime.strptime(start_date_str, "%d.%m.%Y %H:%M:%S")
-        # Преобразуем в UTC с учётом часового пояса
-        timestamp = int(start_date.timestamp()) - (timezone * 3600)  # Сдвигаем в UTC
-        current_time = int(datetime.now().timestamp())  # Текущий UTC timestamp
+        timestamp = int(start_date.timestamp()) - (timezone * 3600)
+        current_time = int(datetime.now().timestamp())
 
-        # Проверяем, не в прошлом ли дата
         if timestamp < current_time:
             await message.answer(
                 text="Ошибка: указанная дата и время находятся в прошлом! Введите будущую дату:",
@@ -225,9 +228,8 @@ async def process_start_date(message: Message, state: FSMContext):
     await state.set_state(RecordingCreateStates.waiting_for_recurrence)
 
 
-@user.callback_query(F.data == "back_to_timezone", RecordingCreateStates.waiting_for_recurrence)
+@admin.callback_query(F.data == "back_to_timezone", RecordingCreateStates.waiting_for_recurrence)
 async def back_to_timezone_from_recurrence(callback: CallbackQuery, state: FSMContext):
-    """Return to timezone input from recurrence."""
     await callback.message.edit_text(
         text="Укажите тайм-зону относительно UTC (например, +3 или -5):",
         reply_markup=InlineKeyboardMarkup(inline_keyboard=[
@@ -238,9 +240,8 @@ async def back_to_timezone_from_recurrence(callback: CallbackQuery, state: FSMCo
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith("recurrence_"), RecordingCreateStates.waiting_for_recurrence)
+@admin.callback_query(F.data.startswith("recurrence_"), RecordingCreateStates.waiting_for_recurrence)
 async def process_recurrence(callback: CallbackQuery, state: FSMContext):
-    """Process recurrence choice and either ask for periodicity or finish."""
     recurrence = callback.data == "recurrence_yes"
     await state.update_data(recurrence=recurrence)
 
@@ -259,9 +260,8 @@ async def process_recurrence(callback: CallbackQuery, state: FSMContext):
     await callback.answer("")
 
 
-@user.callback_query(F.data == "back_to_date", RecordingCreateStates.waiting_for_periodicity)
+@admin.callback_query(F.data == "back_to_date", RecordingCreateStates.waiting_for_periodicity)
 async def back_to_date_from_periodicity(callback: CallbackQuery, state: FSMContext):
-    """Return to recurrence question from periodicity."""
     await callback.message.edit_text(
         text="Это регулярная встреча?",
         reply_markup=InlineKeyboardMarkup(inline_keyboard=[
@@ -274,17 +274,15 @@ async def back_to_date_from_periodicity(callback: CallbackQuery, state: FSMConte
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith("period_"), RecordingCreateStates.waiting_for_periodicity)
+@admin.callback_query(F.data.startswith("period_"), RecordingCreateStates.waiting_for_periodicity)
 async def process_periodicity(callback: CallbackQuery, state: FSMContext):
-    """Process periodicity choice and finish recording creation."""
-    periodicity = int(callback.data.split("_")[1])  # Extract 1 or 2
+    periodicity = int(callback.data.split("_")[1])
     await state.update_data(periodicity=periodicity)
     await finish_recording(callback, state)
     await callback.answer("")
 
 
 async def finish_recording(callback: CallbackQuery, state: FSMContext):
-    """Finish the recording creation process and save to database."""
     state_data = await state.get_data()
     tag_id = state_data.get("tag_id")
     meet_link = state_data.get("meet_link")
@@ -292,6 +290,11 @@ async def finish_recording(callback: CallbackQuery, state: FSMContext):
     timezone = state_data.get("timezone")
     recurrence = state_data.get("recurrence")
     periodicity = state_data.get("periodicity", None)
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
+
     print("Add conf with values", tag_id, meet_link, timestamp, timezone, recurrence, periodicity)
 
     success, response = await add_conference_to_db(
@@ -306,7 +309,7 @@ async def finish_recording(callback: CallbackQuery, state: FSMContext):
         await callback.message.delete()
         await callback.message.answer(
             text=response,
-            reply_markup=main_actions_keyboard,
+            reply_markup=main_actions_keyboard(user.role),
         )
     else:
         await callback.message.answer(
@@ -316,13 +319,16 @@ async def finish_recording(callback: CallbackQuery, state: FSMContext):
     await state.clear()
 
 
-@user.callback_query(F.data == Callbacks.cancel_primary_action_callback)
+@admin.callback_query(F.data == Callbacks.cancel_primary_action_callback)
 async def on_cancel_primary_callback(callback: CallbackQuery, state: FSMContext):
-    """Handle the cancellation of the current action."""
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     await callback.answer("")
     await callback.message.delete()
     await callback.message.answer(
         text="Действие отменено. Выберите новое действие с помощью кнопок под клавиатурой.",
-        reply_markup=main_actions_keyboard,
+        reply_markup=main_actions_keyboard(user.role),
     )
     await state.clear()
