+++ app/roles/user/main_actions/recording_search/recording_search.py
@@ -6,34 +6,35 @@ from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import CallbackQuery, Message, InlineKeyboardButton, InlineKeyboardMarkup
 
 from app.config import labels
+from app.config.roles import Role
 from app.database.conference_db_operations import (
     get_conferences_by_tag,
     get_conference_by_link,
     get_conference_by_id,
-    delete_conference_by_id,  # Добавлен импорт
 )
 from app.database.recording_db_operations import get_recording_by_id
 from app.database.tag_db_operations import get_tag_by_id
+from app.database.user_db_operations import get_user_by_telegram_tag
 from app.keyboards import (
     choose_recordings_search_method_keyboard as recordings_keyboard,
     inline_active_tag_list,
     inline_single_cancel_button,
     main_actions_keyboard,
 )
+from app.roles.admin.admin import admin  # Импортируем admin для удаления
 from app.roles.user.callbacks_enum import Callbacks
-from app.utils.logger import logger
 from app.roles.user.user_cmds import user
+from app.utils.logger import logger
 
 
 class RecordingSearchStates(StatesGroup):
     waiting_for_meet_link = State()
     browsing_conference = State()
-    confirming_conference_deletion = State()  # Новое состояние для подтверждения
+    confirming_conference_deletion = State()
 
 
 @user.message(F.text == labels.GET_RECORD)
 async def get_recording(message: Message):
-    """Handle the command to search for recordings."""
     logger.info("get_recordings_call")
     await message.answer(
         text="Как вы хотите найти запись: по тегу или по ссылке?",
@@ -43,7 +44,6 @@ async def get_recording(message: Message):
 
 @user.callback_query(F.data == Callbacks.get_recording_by_tag_callback)
 async def get_recording_by_tag(callback: CallbackQuery):
-    """Start the process of searching recordings by tag."""
     await callback.answer("")
     await callback.message.edit_text(
         text="Выберите нужный тег:",
@@ -58,7 +58,6 @@ async def get_recording_by_tag(callback: CallbackQuery):
 
 @user.callback_query(F.data.startswith(Callbacks.tag_clicked_in_search_mode_callback))
 async def process_tag_selection(callback: CallbackQuery, state: FSMContext):
-    """Process the selection of a tag for searching conferences."""
     try:
         tag_id = callback.data.split(":")[1]
     except IndexError:
@@ -68,6 +67,11 @@ async def process_tag_selection(callback: CallbackQuery, state: FSMContext):
     tag = await get_tag_by_id(tag_id)
     if not tag:
         await callback.answer("Ошибка: тег не найден в базе данных!", show_alert=True)
+        return
+
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
         return
 
     conferences = await get_conferences_by_tag(tag_id)
@@ -107,7 +111,6 @@ async def process_tag_selection(callback: CallbackQuery, state: FSMContext):
 
 @user.callback_query(F.data == Callbacks.get_recording_by_link_callback)
 async def start_recording_by_link(callback: CallbackQuery, state: FSMContext):
-    """Start the process of searching recordings by conference link."""
     await callback.answer("")
     await callback.message.edit_text(
         text="Введите ссылку на Google Meet конференцию:",
@@ -118,8 +121,12 @@ async def start_recording_by_link(callback: CallbackQuery, state: FSMContext):
 
 @user.message(RecordingSearchStates.waiting_for_meet_link)
 async def process_meet_link(message: Message, state: FSMContext):
-    """Process the entered Google Meet link to find the conference and its recordings."""
     meet_link = message.text.strip()
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
+
     conference = await get_conference_by_link(meet_link)
     if conference:
         tag = await get_tag_by_id(str(conference.tag_id))
@@ -143,12 +150,14 @@ async def process_meet_link(message: Message, state: FSMContext):
                     )
         keyboard = InlineKeyboardMarkup(
             inline_keyboard=[[btn] for btn in buttons]) if buttons else InlineKeyboardMarkup(inline_keyboard=[])
-        # Добавляем кнопки "Удалить" и "Назад"
-        keyboard.inline_keyboard.extend([
-            [InlineKeyboardButton(text="🗑️ Удалить",
-                                  callback_data=f"{Callbacks.delete_conference_callback}:{conference.id}")],
+        if user.role >= Role.ADMIN:
+            keyboard.inline_keyboard.extend([
+                [InlineKeyboardButton(text="🗑️ Удалить",
+                                      callback_data=f"{Callbacks.delete_conference_callback}:{conference.id}")],
+            ])
+        keyboard.inline_keyboard.append(
             [InlineKeyboardButton(text="↩ Назад", callback_data=Callbacks.cancel_primary_action_callback)]
-        ])
+        )
         await message.answer(
             text=response.strip(),
             reply_markup=keyboard,
@@ -164,23 +173,30 @@ async def process_meet_link(message: Message, state: FSMContext):
 
 @user.callback_query(F.data == Callbacks.cancel_primary_action_callback)
 async def on_cancel_primary_callback(callback: CallbackQuery, state: FSMContext):
-    """Handle the cancellation of the current action."""
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     await callback.answer("")
     await callback.message.delete()
     await callback.message.answer(
         text="Действие отменено. Выберите новое действие с помощью кнопок под клавиатурой.",
-        reply_markup=main_actions_keyboard,
+        reply_markup=main_actions_keyboard(user.role),
     )
     await state.clear()
 
 
 @user.callback_query(F.data.startswith("open_conference"))
 async def handle_conference_button(callback: CallbackQuery, state: FSMContext):
-    """Handle the click on a conference button to show its recordings."""
     conference_id = callback.data.split(":")[1]
     conference = await get_conference_by_id(conference_id)
     if not conference:
         await callback.answer("Ошибка: конференция не найдена!", show_alert=True)
+        return
+
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
         return
 
     tag = await get_tag_by_id(str(conference.tag_id))
@@ -204,13 +220,15 @@ async def handle_conference_button(callback: CallbackQuery, state: FSMContext):
                 )
     keyboard = InlineKeyboardMarkup(inline_keyboard=[[btn] for btn in buttons]) if buttons else InlineKeyboardMarkup(
         inline_keyboard=[])
-    # Добавляем кнопки "Удалить" и "Назад"
-    keyboard.inline_keyboard.extend([
-        [InlineKeyboardButton(text="🗑️ Удалить",
-                              callback_data=f"{Callbacks.delete_conference_callback}:{conference.id}")],
+    if user.role >= Role.ADMIN:
+        keyboard.inline_keyboard.extend([
+            [InlineKeyboardButton(text="🗑️ Удалить",
+                                  callback_data=f"{Callbacks.delete_conference_callback}:{conference.id}")],
+        ])
+    keyboard.inline_keyboard.append(
         [InlineKeyboardButton(text="↩ Назад",
                               callback_data=f"{Callbacks.back_to_tag_in_search_mode}:{conference.tag_id}")]
-    ])
+    )
 
     await callback.message.edit_text(
         text=response.strip(),
@@ -222,11 +240,15 @@ async def handle_conference_button(callback: CallbackQuery, state: FSMContext):
 
 @user.callback_query(F.data.startswith(Callbacks.back_to_tag_in_search_mode))
 async def handle_back_to_tag_in_search_mode(callback: CallbackQuery, state: FSMContext):
-    """Handle the 'Back' button to return to the list of conferences by tag."""
     tag_id = callback.data.split(":")[1]
     tag = await get_tag_by_id(tag_id)
     if not tag:
         await callback.answer("Ошибка: тег не найден!", show_alert=True)
+        return
+
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
         return
 
     conferences = await get_conferences_by_tag(tag_id)
@@ -264,9 +286,8 @@ async def handle_back_to_tag_in_search_mode(callback: CallbackQuery, state: FSMC
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith(Callbacks.delete_conference_callback))
+@admin.callback_query(F.data.startswith(Callbacks.delete_conference_callback))
 async def handle_delete_conference(callback: CallbackQuery, state: FSMContext):
-    """Handle the 'Delete' button to confirm conference deletion."""
     conference_id = callback.data.split(":")[1]
     conference = await get_conference_by_id(conference_id)
     if not conference:
@@ -287,34 +308,42 @@ async def handle_delete_conference(callback: CallbackQuery, state: FSMContext):
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith(Callbacks.confirm_delete_conference))
+@admin.callback_query(F.data.startswith(Callbacks.confirm_delete_conference))
 async def confirm_delete_conference(callback: CallbackQuery, state: FSMContext):
-    """Handle confirmation of conference deletion."""
+    from app.database.conference_db_operations import delete_conference_by_id  # Импорт внутри функции
     conference_id = callback.data.split(":")[1]
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     success, response = await delete_conference_by_id(conference_id)
     if success:
         await callback.message.delete()
         await callback.message.answer(
             text=response,
-            reply_markup=main_actions_keyboard,
+            reply_markup=main_actions_keyboard(user.role),
         )
     else:
         await callback.message.answer(
             text=response,
-            reply_markup=main_actions_keyboard,
+            reply_markup=main_actions_keyboard(user.role),
         )
     await state.clear()
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith(Callbacks.cancel_delete_conference))
+@admin.callback_query(F.data.startswith(Callbacks.cancel_delete_conference))
 async def cancel_delete_conference(callback: CallbackQuery, state: FSMContext):
-    """Handle cancellation of conference deletion."""
     conference_id = callback.data.split(":")[1]
     conference = await get_conference_by_id(conference_id)
     if not conference:
         await callback.answer("Ошибка: конференция не найдена!", show_alert=True)
         await state.clear()
+        return
+
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
         return
 
     tag = await get_tag_by_id(str(conference.tag_id))
@@ -338,12 +367,15 @@ async def cancel_delete_conference(callback: CallbackQuery, state: FSMContext):
                 )
     keyboard = InlineKeyboardMarkup(inline_keyboard=[[btn] for btn in buttons]) if buttons else InlineKeyboardMarkup(
         inline_keyboard=[])
-    keyboard.inline_keyboard.extend([
-        [InlineKeyboardButton(text="🗑️ Удалить",
-                              callback_data=f"{Callbacks.delete_conference_callback}:{conference.id}")],
+    if user.role >= Role.ADMIN:
+        keyboard.inline_keyboard.extend([
+            [InlineKeyboardButton(text="🗑️ Удалить",
+                                  callback_data=f"{Callbacks.delete_conference_callback}:{conference.id}")],
+        ])
+    keyboard.inline_keyboard.append(
         [InlineKeyboardButton(text="↩ Назад",
                               callback_data=f"{Callbacks.back_to_tag_in_search_mode}:{conference.tag_id}")]
-    ])
+    )
 
     await callback.message.edit_text(
         text=response.strip(),
